PythonCode/CNNIQA/requirements.txt
argparse==1.4.0
h5py~=3.9.0
PyYAML~=6.0
Pillow~=9.5.0
numpy~=1.22.4
scipy~=1.10.1
torch~=1.11.0
torchvision~=0.15.2
pytorch-ignite==0.2.1
tensorflow-gpu==2.2.0
tensorboardX~=2.6

yaml~=0.2.5
matplotlib~=3.7.1
scikit-image~=0.22.0
pandas~=2.0.2
ignite~=1.1.0
linformer~=0.2.1
scikit-learn~=1.2.2
tqdm~=4.65.0
tensorflow~=2.12.0
einops~=0.6.1
openpyxl~=3.1.2
seaborn~=0.12.2